requirements.txt
asgiref==3.7.1
Django==4.2.1
djangorestframework==3.14.0
geographiclib==2.0
geopy==2.3.0
pip==22.0.4
psycopg2==2.9.6
pytz==2023.3
setuptools==58.1.0
sqlparse==0.4.4
typing_extensions==4.6.1
tzdata==2023.3